requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioquic==1.2.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.8.1
attrs==25.3.0
Automat==25.4.16
bleach==6.2.0
blinker==1.9.0
bokeh==3.7.3
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
constantly==23.10.4
contourpy==1.3.3
cryptography==44.0.3
distro==1.9.0
Flask==3.1.0
frozenlist==1.7.0
geoip2==5.1.0
h11==0.16.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
kaitaistruct==0.10
ldap3==2.9.1
libvirt-python==11.6.0
linkify-it-py==2.0.3
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
maxminddb==2.8.2
mdit-py-plugins==0.5.0
mdurl==0.1.2
mitmproxy==12.1.1
mitmproxy_linux==0.12.7
mitmproxy_rs==0.12.7
msgpack==1.1.0
multidict==6.6.4
narwhals==2.1.0
numpy==2.3.2
openai==2.8.1
packaging==25.0
pandas==2.3.1
panel==1.7.5
param==2.2.1
passlib==1.7.4
pillow==11.3.0
propcache==0.3.2
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pylsqpack==0.3.22
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.2
requests==2.32.4
ruamel.yaml==0.18.10
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tornado==6.5
tqdm==4.67.1
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
urwid==2.6.16
vncdotool==1.2.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
xyzservices==2025.4.0
yarl==1.20.1
zope.interface==7.2
zstandard==0.23.0
